Refactoring advancement handler/helper
