+++ module/documents/actor.mjs
@@ -1,8 +1,8 @@
 import { Dice } from "../dice.mjs";
 import { RollDialog } from "../helpers/roll-dialog.mjs";
 import { resizeTokens } from "../helpers/actor.mjs";
-import { roleValueChange } from "../helpers/advancement.mjs";
 import { getItemsOfType } from "../helpers/utils.mjs";
+import { roleValueChange } from "../sheet-handlers/role-handler.mjs";
 
 /**
  * Extend the base Actor document by defining a custom roll data structure which is ideal for the Simple system.
